pyserver/requirements.txt
obswebsocket
flask
flask_cors
mss
pillow
io
json
os